refactor(frontend): split Layout god file into PublicLayout + AdminLayout
Layout.tsx becomes an 18-line router switch; public header/nav
extracted to PublicLayout.tsx, admin sidebar/nav to AdminLayout.tsx,
BrandFallbackTile extracted to its own file for reuse.
Public API of Layout unchanged; App.tsx import untouched.
